requirements.txt
pandas
numpy
statsmodels
torch
streamlit
scipy
python-dotenv
groq
